requirements.txt
torch
torchvision
torchaudio
numpy
pandas
scikit-learn
matplotlib
pillow
tqdm
pyyaml
tensorboard
torchmetrics
jupyter
streamlit>=1.28.0
streamlit-autorefresh>=1.0.1
plotly>=5.15.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pyserial>=3.5
